refactor[#45] : Menu Icon 오탈자 수정
