+++ beta-reader/src/shared/assets/MenuIcon.tsx
